wip: refactor: TaskView skeleton
