+++ tedi_ops/src/lib.rs
@@ -26,8 +26,8 @@ pub use tedi_core::current_user;
 // Data model re-exported at the crate root so moved code keeps resolving `crate::X`.
 pub use tedi_core::{
 	BlockerItem, BlockerSetState, Blockers, CloseState, Comment, CommentIdentity, Comments, HollowIssue, Issue, IssueContents, IssueError, IssueIdentity, IssueIndex, IssueLink, IssueMarker,
-	IssueRef, IssueSelector, IssueTimestamps, LinkedIssueMeta, MAX_INDEX_DEPTH, MAX_LINEAGE_DEPTH, MAX_TITLE_LENGTH, Marker, Milestone, ParseError, RepoInfo, TitleInGitPathError,
-	VirtualIssue, parse_blockers_from_embedded, split_blockers,
+	IssueRef, IssueSelector, IssueTimestamps, LinkedIssueMeta, MAX_INDEX_DEPTH, MAX_LINEAGE_DEPTH, MAX_TITLE_LENGTH, Marker, MilestoneLink, MilestoneRef, ParseError, RepoInfo, TaskView,
+	TitleInGitPathError, VirtualIssue, parse_blockers_from_embedded, split_blockers,
 };
 pub use tedi_md::{Events, Header, OwnedCodeBlockKind, OwnedEvent, OwnedTag, OwnedTagEnd};
